Jorrit fixed a 64-bit issue in AWS: Broadcast was using an unsigned long. Should be an uint32. unsigned long caused signalComponentCreated messages to fail.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22920 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/iaws/aws.h
@@ -724,7 +724,7 @@ struct iAwsSource : public iBase
   virtual bool UnregisterSlot(iAwsSlot *slot, unsigned long signal) = 0;
 
   /// Broadcasts a signal to all slots that are interested.
-  virtual void Broadcast(unsigned long signal) = 0;
+  virtual void Broadcast(uint32 signal) = 0;
 };
