+++ plugins/aws/awsslot.h
@@ -201,7 +201,7 @@ public:
   virtual bool UnregisterSlot (iAwsSlot *slot, unsigned long signal);
 
   /// Broadcasts a signal to all slots that are interested.
-  virtual void Broadcast (unsigned long signal);
+  virtual void Broadcast (uint32 signal);
 };
 
 /**
